requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
pydantic==2.9.0
psutil==6.0.0
pyotp==2.9.0
camoufox[geoip]==0.4.11
playwright==1.48.0
python-dotenv==1.0.1
